Iter factor in tabu search now accepted float arguments

+++ src/ts/ts.h
@@ -20,7 +20,7 @@ struct TSResult {
 
 TSResult tspmed (const Instance&    ,
                  const vector <int>&,
-                 int = 2,
+                 double = 2.0,
                  shared_ptr <SolutionTrie> = nullptr);
 
 #endif
